refactor: extract shared domain types into model.rs
VarScope, EnvValue, EnvVar, EnvSnapshot, EnvChange, and
UnsupportedEnvValue were defined in env_store.rs but imported
directly by nine other files (env_backend, export/*, commands/*,
snapshot, cli) that have nothing to do with env_store's read/write/
apply logic. Moving them to a dedicated model.rs makes those imports
reflect what's actually being borrowed: shared data shapes, not
env_store's business logic.
