+++ src-tauri/src/export/mod.rs
@@ -1,5 +1,5 @@
-use crate::env_store::{EnvSnapshot, EnvValueKind};
 use crate::error::EnvarlyError;
+use crate::model::{EnvSnapshot, EnvValueKind};
 
 mod ansible;
 mod dsc;
